# docs/swagger.yaml
basePath: /api
definitions:
...
  core.JSONB:
    additionalProperties: true
    type: object
...
  credit.CreditNote:
    properties:
      amount:
        description: The integer amount in %s representing the total amount of the
          credit note, including tax.
        type: integer
      created_at:
        type: string
      currency:
        description: Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html),
          in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).
        type: string
      customer:
        format: uuid
        type: string
      customer_balance_transaction:
        format: uuid
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      discount_amount:
        description: The integer amount in %s representing the total amount of discount
          that was credited.
        type: integer
      id:
        type: string
      invoice:
        format: uuid
        type: string
      lines:
        description: The aggregate amounts calculated per discount for all line items.
        items:
          format: uuid
          type: string
        type: array
      livemode:
        type: boolean
      memo:
        description: Customer-facing text that appears on the credit note PDF.
        type: string
      metadata:
        $ref: '#/definitions/core.JSONB'
      number:
        description: A unique number that identifies this particular credit note and
          appears on the PDF of the credit note and its associated invoice.
        type: string
      object:
        type: string
      out_of_band_amount:
        description: Amount that was credited outside of Stripe.
        type: integer
      pdf:
        description: The link to download the PDF of the credit note.
        type: string
      reason:
        description: Reason for issuing this credit note, one of `duplicate`, `fraudulent`,
          `order_change`, or `product_unsatisfactory`
        format: uuid
        type: string
      refund:
        format: uuid
        type: string
      status:
        description: Status of this credit note, one of `issued` or `void`. Learn
          more about [voiding credit notes](https://stripe.com/docs/billing/invoices/credit-notes#voiding).
        format: uuid
        type: string
      subtotal:
        description: The integer amount in %s representing the amount of the credit
          note, excluding exclusive tax and invoice level discounts.
        type: integer
      subtotal_excluding_tax:
        description: The integer amount in %s representing the amount of the credit
          note, excluding all tax and invoice level discounts.
        type: integer
      tax_amounts:
        description: The aggregate amounts calculated per tax rate for all line items.
        items:
          format: uuid
          type: string
        type: array
      total:
        description: The integer amount in %s representing the total amount of the
          credit note, including tax and all discount.
        type: integer
      total_excluding_tax:
        description: The integer amount in %s representing the total amount of the
          credit note, excluding tax, but including discounts.
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/credit.Type'
        description: Type of this credit note, one of `pre_payment` or `post_payment`.
          A `pre_payment` credit note means it was issued when the invoice was open.
          A `post_payment` credit note means it was issued when the invoice was paid.
      updated_at:
        type: string
      voided_at:
        description: The time that the credit note was voided.
        type: string
    type: object
  credit.Type:
    enum:
    - pre_payment
    - post_payment
    type: string
    x-enum-varnames:
    - TypePrePayment
    - TypePostPayment
...
  database.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  item.LineItem:
    properties:
      amount:
        type: integer
      amount_excluding_tax:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      description:
        type: string
      discountable:
        type: boolean
      discounts:
        items:
          format: uuid
          type: string
        type: array
      id:
        type: string
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      object:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      price:
        format: uuid
        type: string
      product:
        format: uuid
        type: string
      proration:
        type: boolean
      proration_details:
        format: uuid
        type: string
      quantity:
        type: integer
      subscription:
        type: string
      subscription_item:
        type: string
      tax_rates:
        items:
          format: uuid
          type: string
        type: array
      type:
        type: string
      unit_amount_excluding_tax:
        type: number
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /v1/credit_notes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete LineItem by given ID.
      parameters:
      - description: LineItem ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete LineItem by given ID
      tags:
      - LineItem
    get:
      consumes:
      - application/json
      description: Get LineItem by given ID or 404 error.
      parameters:
      - description: LineItem ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.LineItem'
      summary: get LineItem by given ID or 404 error.
      tags:
      - LineItem
    post:
      consumes:
      - application/json
      description: Update LineItem.
      parameters:
      - description: LineItem ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.LineItem'
      summary: update LineItem
      tags:
      - LineItem
  /v1/credit_notes/{id}/lines:
    get:
      consumes:
      - application/json
      description: Get CreditNote by given ID and return his Linitems.
      parameters:
      - description: CreditNote ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit.CreditNote'
      summary: get CreditNote by given ID and return his Linitems.
      tags:
      - CreditNote
  /v1/credit_notes/{id}/void:
    get:
      consumes:
      - application/json
      description: Update CreditNote to status void.
      parameters:
      - description: CreditNote ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit.CreditNote'
      summary: update CreditNote to status void
      tags:
      - CreditNote
...
schemes:
- http
swagger: "2.0"